[ref] root perm

+++ install.sh
@@ -1,16 +1,19 @@
+echo "We need root permissions to move the executable to /usr/sbin"
+
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as root"
+  exit
+fi
+
 PASS=$HOME/.go-pass
 
-# rm -rf $PASS
+rm -rf $PASS
 
 mkdir $PASS &>/dev/null;
 
-rm -rf $PASS;
-
 git clone https://github.com/victorlpgazolli/go-pass $PASS;
 
-echo "We need sudo permissions to move the executable to /usr/sbin"
-
-sudo cp $PASS/pass /usr/sbin/pass
+cp $PASS/pass /usr/sbin/pass
 
 echo "pass installed in /usr/sbin"
